backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
google-genai>=0.5.0
langgraph>=0.0.30
langgraph-checkpoint-sqlite>=0.0.1
pydantic>=2.6.0
python-dotenv>=1.0.1
pyyaml>=6.0.1
tree-sitter>=0.21.3
tree-sitter-python>=0.21.0
tree-sitter-javascript>=0.21.0
tree-sitter-typescript>=0.21.0
psutil>=5.9.0
chromadb>=0.4.24
pytest>=8.0.0
llama-parse>=0.4.0
aiosqlite>=0.20.0
watchdog>=4.0.0
lupa>=2.2
libcst>=1.0.0

